refactor: UiState 에 있는 로직들 ViewModel 로 옮기기
