+++ app/src/main/java/woowacourse/shopping/presentation/cart/CartViewModel.kt
@@ -154,3 +154,112 @@ class CartViewModel(
         }
     }
 }
+
+private fun CartUiState.toggleProductSelected(productId: Long): CartUiState {
+    val newCurrentProducts =
+        currentPageProducts.map {
+            if (it.product.id == productId) {
+                it.copy(isSelected = !it.isSelected)
+            } else {
+                it
+            }
+        }
+    val newPagingProducts = pagingProducts + (currentPage to newCurrentProducts)
+    return copy(
+        pagingProducts = newPagingProducts,
+    )
+}
+
+private fun CartUiState.toggleTotalOrderProducts(): CartUiState {
+    val isSelectedAll = isTotalProductsOrdered
+    val newPagingProducts =
+        pagingProducts.map {
+            it.key to
+                    it.value.map { cartProductUi ->
+                        cartProductUi.copy(isSelected = !isSelectedAll)
+                    }
+        }.toMap()
+    return copy(
+        pagingProducts = newPagingProducts,
+    )
+}
+
+private fun CartUiState.updateTotalProducts(products: List<CartProductUi>): CartUiState {
+    val newPagingProducts =
+        products.chunked(5).mapIndexed { index, chunkedProducts ->
+            (index + 1) to chunkedProducts
+        }.toMap()
+    return copy(
+        pagingProducts = newPagingProducts,
+    )
+}
+
+private fun CartUiState.updateTargetPageProducts(
+    products: List<CartProductUi>,
+    targetPage: Int,
+): CartUiState {
+    val originalCurrentPageProducts = pagingProducts[targetPage] ?: products
+    val ids = originalCurrentPageProducts.map { it.product.id }
+    val newProducts =
+        products.map { cartProduct ->
+            val newId = cartProduct.product.id
+            if (newId in ids) {
+                val original = originalCurrentPageProducts.find { it.product.id == newId }
+                return@map original ?: cartProduct
+            }
+            cartProduct
+        }
+
+    val newPagingProducts = pagingProducts + (targetPage to newProducts)
+    return copy(
+        currentPage = targetPage,
+        pagingProducts = newPagingProducts,
+    )
+}
+
+private fun CartUiState.increaseProductCount(
+    productId: Long,
+    amount: Int,
+): CartUiState {
+    val newCurrentProducts =
+        currentPageProducts.map {
+            if (it.product.id == productId) {
+                it.copy(count = it.count + amount)
+            } else {
+                it
+            }
+        }
+    val newPagingProducts = pagingProducts + (currentPage to newCurrentProducts)
+    return copy(
+        pagingProducts = newPagingProducts,
+    )
+}
+
+private fun CartUiState.decreaseProductCount(
+    productId: Long,
+    amount: Int,
+): CartUiState {
+    val newCurrentProducts =
+        currentPageProducts.map {
+            if (it.product.id == productId) {
+                it.copy(count = it.count - amount)
+            } else {
+                it
+            }
+        }
+    val newPagingProducts = pagingProducts + (currentPage to newCurrentProducts)
+    return copy(
+        pagingProducts = newPagingProducts,
+    )
+}
+
+private fun CartUiState.canDecreaseProductCount(
+    productId: Long,
+    countLimit: Int,
+): Boolean {
+    val product = findProductAtCurrentPage(productId) ?: return false
+    return product.count > countLimit
+}
+
+private fun CartUiState.findProductAtCurrentPage(productId: Long): CartProductUi? =
+    currentPageProducts.find { it.product.id == productId }
