+++ app/src/main/java/woowacourse/shopping/presentation/cart/recommend/RecommendOrderUiState.kt
@@ -15,44 +15,6 @@ data class RecommendOrderUiState(
         get() = totalProducts.sumOf { it.count }
 
     val totalPrice
-        get() =
-            totalProducts
-                .sumOf { it.totalPrice }
-
-    fun increaseProductCount(
-        productId: Long,
-        amount: Int,
-    ): RecommendOrderUiState =
-        copy(
-            recommendProducts =
-                recommendProducts.map {
-                    if (it.product.id == productId) {
-                        it.copy(count = it.count + amount)
-                    } else {
-                        it
-                    }
-                },
-        )
-
-    fun decreaseProductCount(
-        productId: Long,
-        amount: Int,
-    ): RecommendOrderUiState {
-        val newProducts =
-            recommendProducts.map {
-                if (it.product.id == productId) {
-                    it.copy(count = it.count - amount)
-                } else {
-                    it
-                }
-            }
-        return copy(recommendProducts = newProducts)
-    }
-
-    fun shouldDeleteFromCart(productId: Long): Boolean {
-        val product = findProduct(productId) ?: return false
-        return product.count <= 1
-    }
-
-    fun findProduct(productId: Long): CartProductUi? = recommendProducts.find { it.product.id == productId }
+        get() = totalProducts
+            .sumOf { it.totalPrice }
 }
